modified Ethernet class

+++ src/Router.java
@@ -271,7 +271,7 @@ public class Router {
 			System.out.println("ip3 = ip3 " + ip3.equals(ip3));
 			
 			print("\nmaking ethernet frame\n");
-			EthernetFrame ef1 = new EthernetFrame(m3,m2,255,new byte[]{'a','b','c'});
+			EthernetFrame ef1 = new EthernetFrame(m3,m2,(short)255,new byte[]{'a','b','c'});
 			System.out.println(ef1.toString());
 			
 //			print("test crc32 class\n");
